+++ embassy-nrf/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- next-header -->
 ## Unreleased - ReleaseDate
+- changed: apply trimming values from FICR.TRIMCNF on nrf53/54l
 
 ## 0.8.0 - 2025-09-30